+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncSortedSet.java
@@ -33,6 +33,7 @@ import com.github.jlangch.venice.impl.types.Constants;
 import com.github.jlangch.venice.impl.types.TypeRank;
 import com.github.jlangch.venice.impl.types.VncKeyword;
 import com.github.jlangch.venice.impl.types.VncVal;
+import com.github.jlangch.venice.impl.types.custom.VncWrappingTypeDef;
 import com.github.jlangch.venice.impl.types.util.Types;
 import com.github.jlangch.venice.impl.util.EmptyIterator;
 import com.github.jlangch.venice.impl.util.MetaUtil;
@@ -41,11 +42,18 @@ import com.github.jlangch.venice.impl.util.MetaUtil;
 public class VncSortedSet extends VncSet {
 
     public VncSortedSet() {
-        this(null, null);
+        this((io.vavr.collection.Set<VncVal>)null, null);
     }
 
     public VncSortedSet(final VncVal meta) {
-        this(null, meta);
+        this((io.vavr.collection.Set<VncVal>)null, meta);
+    }
+
+    public VncSortedSet(
+            final VncWrappingTypeDef wrappingTypeDef,
+            final VncVal meta
+    ) {
+        this(null, wrappingTypeDef, meta);
     }
 
     public VncSortedSet(final io.vavr.collection.Set<VncVal> val) {
@@ -53,7 +61,7 @@ public class VncSortedSet extends VncSet {
     }
 
     public VncSortedSet(final io.vavr.collection.Set<VncVal> val, final VncVal meta) {
-        super(meta == null ? Constants.Nil : meta);
+        super(meta);
         if (val == null) {
             value = io.vavr.collection.TreeSet.empty();
         }
@@ -65,6 +73,22 @@ public class VncSortedSet extends VncSet {
         }
     }
 
+    public VncSortedSet(
+            final io.vavr.collection.Set<VncVal> val,
+            final VncWrappingTypeDef wrappingTypeDef,
+            final VncVal meta
+    ) {
+        super(wrappingTypeDef, meta);
+        if (val == null) {
+            value = io.vavr.collection.TreeSet.empty();
+        }
+        else if (val instanceof io.vavr.collection.TreeSet) {
+            value = (io.vavr.collection.TreeSet<VncVal>)val;
+        }
+        else {
+            value = io.vavr.collection.TreeSet.ofAll(val);
+        }
+    }
 
     public static VncSortedSet ofAll(final java.util.Collection<? extends VncVal> val) {
         return new VncSortedSet(io.vavr.collection.TreeSet.ofAll(val));
@@ -100,13 +124,25 @@ public class VncSortedSet extends VncSet {
     }
 
     @Override
+    public VncSortedSet wrap(final VncWrappingTypeDef wrappingTypeDef, final VncVal meta) {
+        return new VncSortedSet(value, wrappingTypeDef, meta);
+    }
+
+    @Override
     public VncKeyword getType() {
-        return new VncKeyword(
-                        TYPE,
-                        MetaUtil.typeMeta(
-                            new VncKeyword(VncSet.TYPE),
-                            new VncKeyword(VncCollection.TYPE),
-                            new VncKeyword(VncVal.TYPE)));
+        return isWrapped() ? new VncKeyword(
+                                    getWrappingTypeDef().getType().getQualifiedName(),
+                                    MetaUtil.typeMeta(
+                                        new VncKeyword(VncSortedSet.TYPE),
+                                        new VncKeyword(VncSet.TYPE),
+                                        new VncKeyword(VncCollection.TYPE),
+                                        new VncKeyword(VncVal.TYPE)))
+                           : new VncKeyword(
+                                    VncSortedSet.TYPE,
+                                    MetaUtil.typeMeta(
+                                            new VncKeyword(VncSet.TYPE),
+                                            new VncKeyword(VncCollection.TYPE),
+                                            new VncKeyword(VncVal.TYPE)));
     }
 
     @Override