+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncVector.java
@@ -38,6 +38,7 @@ import com.github.jlangch.venice.impl.types.TypeRank;
 import com.github.jlangch.venice.impl.types.VncKeyword;
 import com.github.jlangch.venice.impl.types.VncString;
 import com.github.jlangch.venice.impl.types.VncVal;
+import com.github.jlangch.venice.impl.types.custom.VncWrappingTypeDef;
 import com.github.jlangch.venice.impl.types.util.Coerce;
 import com.github.jlangch.venice.impl.types.util.Types;
 import com.github.jlangch.venice.impl.util.ArityExceptions;
@@ -50,6 +51,10 @@ public class VncVector extends VncSequence implements IVncFunction {
 
     protected VncVector(final VncVal meta) {
         this((io.vavr.collection.Seq<VncVal>)null, meta);
+    }
+
+    protected VncVector(final VncWrappingTypeDef wrappingTypeDef, final VncVal meta) {
+        this((io.vavr.collection.Seq<VncVal>)null, wrappingTypeDef, meta);
     }
 
     protected VncVector(final java.util.Collection<? extends VncVal> vals, final VncVal meta) {
@@ -68,6 +73,24 @@ public class VncVector extends VncSequence implements IVncFunction {
             value = io.vavr.collection.Vector.ofAll(vals);
         }
     }
+
+    public VncVector(
+            final io.vavr.collection.Seq<VncVal> vals,
+            final VncWrappingTypeDef wrappingTypeDef,
+            final VncVal meta
+    ) {
+        super(wrappingTypeDef, meta);
+        if (vals == null) {
+            value = io.vavr.collection.Vector.empty();
+        }
+        else if (vals instanceof io.vavr.collection.Vector) {
+            value = (io.vavr.collection.Vector<VncVal>)vals;
+        }
+        else {
+            value = io.vavr.collection.Vector.ofAll(vals);
+        }
+    }
+
     public static VncVector of(final VncVal... mvs) {
         return mvs.length <= VncTinyVector.MAX_ELEMENTS
                 ? VncTinyVector.of(mvs)
@@ -150,13 +173,26 @@ public class VncVector extends VncSequence implements IVncFunction {
     }
 
     @Override
+    public VncVector wrap(final VncWrappingTypeDef wrappingTypeDef, final VncVal meta) {
+        return new VncVector(value, wrappingTypeDef, meta);
+    }
+
+
+    @Override
     public VncKeyword getType() {
-        return new VncKeyword(
-                        TYPE,
-                        MetaUtil.typeMeta(
-                            new VncKeyword(VncSequence.TYPE),
-                            new VncKeyword(VncCollection.TYPE),
-                            new VncKeyword(VncVal.TYPE)));
+        return isWrapped() ? new VncKeyword(
+                                    getWrappingTypeDef().getType().getQualifiedName(),
+                                    MetaUtil.typeMeta(
+                                        new VncKeyword(VncVector.TYPE),
+                                        new VncKeyword(VncSequence.TYPE),
+                                        new VncKeyword(VncCollection.TYPE),
+                                        new VncKeyword(VncVal.TYPE)))
+                           : new VncKeyword(
+                                    VncVector.TYPE,
+                                    MetaUtil.typeMeta(
+                                            new VncKeyword(VncSequence.TYPE),
+                                            new VncKeyword(VncCollection.TYPE),
+                                            new VncKeyword(VncVal.TYPE)));
     }
 
     @Override