+++ ChangeLog.md
@@ -10,8 +10,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- ... 
-
+- custom type support for Venice persistent collections
 
 ### Fixed